+++ src/Regressors/MLPRegressor.php
@@ -86,13 +86,6 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
     protected \Rubix\ML\NeuralNet\Optimizers\Optimizer $optimizer;
 
     /**
-     * The amount of L2 regularization applied to the weights of the output layer.
-     *
-     * @var float
-     */
-    protected float $l2Penalty;
-
-    /**
      * The maximum number of training epochs. i.e. the number of times to iterate before terminating.
      *
      * @var int<0,max>
@@ -105,6 +98,13 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
      * @var float
      */
     protected float $minChange;
+
+    /**
+     * The number of epochs to train before evaluating the model with the holdout set.
+     *
+     * @var int
+     */
+    protected $evalInterval;
 
     /**
      * The number of epochs without improvement in the validation score to wait before considering an early stop.
@@ -159,9 +159,9 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
      * @param \Rubix\ML\NeuralNet\Layers\Hidden[] $hiddenLayers
      * @param int $batchSize
      * @param \Rubix\ML\NeuralNet\Optimizers\Optimizer|null $optimizer
-     * @param float $l2Penalty
      * @param int $epochs
      * @param float $minChange
+     * @param int $evalInterval
      * @param int $window
      * @param float $holdOut
      * @param \Rubix\ML\NeuralNet\CostFunctions\RegressionLoss|null $costFn
@@ -172,9 +172,9 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
         array $hiddenLayers = [],
         int $batchSize = 128,
         ?Optimizer $optimizer = null,
-        float $l2Penalty = 1e-4,
         int $epochs = 1000,
         float $minChange = 1e-4,
+        int $evalInterval = 3,
         int $window = 5,
         float $holdOut = 0.1,
         ?RegressionLoss $costFn = null,
@@ -192,11 +192,6 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
                 . " greater than 0, $batchSize given.");
         }
 
-        if ($l2Penalty < 0.0) {
-            throw new InvalidArgumentException('L2 Penalty must be'
-                . " greater than 0, $l2Penalty given.");
-        }
-
         if ($epochs < 0) {
             throw new InvalidArgumentException('Number of epochs'
                 . " must be greater than 0, $epochs given.");
@@ -205,6 +200,11 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
         if ($minChange < 0.0) {
             throw new InvalidArgumentException('Minimum change must be'
                 . " greater than 0, $minChange given.");
+        }
+
+        if ($evalInterval < 1) {
+            throw new InvalidArgumentException('Eval interval must be'
+                . " greater than 0, $evalInterval given.");
         }
 
         if ($window < 1) {
@@ -224,9 +224,9 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
         $this->hiddenLayers = $hiddenLayers;
         $this->batchSize = $batchSize;
         $this->optimizer = $optimizer ?? new Adam();
-        $this->l2Penalty = $l2Penalty;
         $this->epochs = $epochs;
         $this->minChange = $minChange;
+        $this->evalInterval = $evalInterval;
         $this->window = $window;
         $this->holdOut = $holdOut;
         $this->costFn = $costFn ?? new LeastSquares();
@@ -272,9 +272,9 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
             'hidden layers' => $this->hiddenLayers,
             'batch size' => $this->batchSize,
             'optimizer' => $this->optimizer,
-            'l2 penalty' => $this->l2Penalty,
             'epochs' => $this->epochs,
             'min change' => $this->minChange,
+            'eval interval' => $this->evalInterval,
             'window' => $this->window,
             'hold out' => $this->holdOut,
             'cost fn' => $this->costFn,
@@ -353,7 +353,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
 
         $hiddenLayers = $this->hiddenLayers;
 
-        $hiddenLayers[] = new Dense(1, $this->l2Penalty, true, new Xavier2());
+        $hiddenLayers[] = new Dense(1, 0.0, true, new Xavier2());
 
         $this->network = new FeedForward(
             new Placeholder1D($dataset->numFeatures()),
@@ -428,7 +428,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
                 break;
             }
 
-            if (!$testing->empty()) {
+            if ($epoch % $this->evalInterval === 0 && !$testing->empty()) {
                 $predictions = $this->predict($testing);
 
                 $score = $this->metric->score($predictions, $testing->labels());
@@ -437,12 +437,11 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
             }
 
             if ($this->logger) {
-                $lossDirection = $loss < $prevLoss ? '↓' : '↑';
+                $message = "Epoch: $epoch, {$this->costFn}: $loss";
 
-                $message = "Epoch: $epoch, "
-                    . "{$this->costFn}: $loss, "
-                    . "Loss Change: {$lossDirection}{$lossChange}, "
-                    . "{$this->metric}: " . ($score ?? 'N/A');
+                if (isset($score)) {
+                    $message .= ", {$this->metric}: $score";
+                }
 
                 $this->logger->info($message);
             }
@@ -466,6 +465,8 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
                 if ($numWorseEpochs >= $this->window) {
                     break;
                 }
+
+                unset($score);
             }
 
             if ($lossChange < $this->minChange) {